+++ packages/assessment-color-shapes/src/index.ts
@@ -147,6 +147,12 @@ class ColorShapes extends Game {
         description:
           "ISO 8601 timestamp at the beginning of the trial. Null if trial was skipped.",
       },
+      trial_end_iso8601_timestamp: {
+        type: ["string", "null"],
+        format: "date-time",
+        description:
+          "ISO 8601 timestamp at the end of the trial (when user presses 'Same' or 'Different'). Null if trial was skipped.",
+      },
       trial_index: {
         type: ["integer", "null"],
         description: "Index of the trial within this assessment, 0-based.",
@@ -818,6 +824,10 @@ phases.`,
       Timer.remove("rt");
       responseGrid.removeAllChildren();
 
+      game.addTrialData(
+        "trial_end_iso8601_timestamp",
+        new Date().toISOString()
+      );
       const trialConfiguration = trialConfigurations[game.trialIndex];
       game.addTrialData("response_time_duration_ms", rt);
       game.addTrialData(